requirements.txt
Flask
requests
groq
python-dotenv